requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
slack-sdk==3.20.2
sqlparse==0.4.3
urllib3==1.26.15
Werkzeug==2.2.3
